requirements.txt
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
biopython==1.79
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.3
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
intervaltree==3.1.0
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.4.1
jupyterlab-pygments==0.2.2
jupyterlab-server==2.13.0
kiwisolver==1.4.2
libcst==0.4.1
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
multiprocess==0.70.12.2
mypy-extensions==0.4.3
nbclassic==0.3.7
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.1
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathos==0.2.8
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pox==0.3.0
ppft==1.6.6.4
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyre-check==0.9.13
pyre-extensions==0.0.27
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
regex==2022.4.24
requests==2.27.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqldiff==0.0.4
sqlparse==0.4.1
stack-data==0.2.0
tabulate==0.8.9
terminado==0.13.3
TestSlide==2.7.0
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.0
typeguard==2.13.3
typing-extensions==4.2.0
typing-inspect==0.7.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
zipp==3.8.0
